+++ exporter/opentelemetry-exporter-otlp-proto-common/tests/test_metrics_encoder.py
@@ -26,6 +26,7 @@ from opentelemetry.sdk.metrics import Exemplar
 from opentelemetry.sdk.metrics.export import (
     AggregationTemporality,
     Buckets,
+    DataPointFlags,
     ExponentialHistogramDataPoint,
     HistogramDataPoint,
     Metric,
@@ -80,6 +81,7 @@ class TestOTLPMetricsEncoder(unittest.TestCase):
                     explicit_bounds=[10.0, 20.0],
                     min=8,
                     max=18,
+                    flags=DataPointFlags.get_default(),
                 )
             ],
             aggregation_temporality=AggregationTemporality.DELTA,
@@ -101,7 +103,13 @@ class TestOTLPMetricsEncoder(unittest.TestCase):
                                 version="first_version",
                                 schema_url="instrumentation_scope_schema_url",
                             ),
-                            metrics=[_generate_sum("sum_int", 33)],
+                            metrics=[
+                                _generate_sum(
+                                    "sum_int",
+                                    33,
+                                    flags=DataPointFlags.get_default(),
+                                )
+                            ],
                             schema_url="instrumentation_scope_schema_url",
                         )
                     ],
@@ -152,6 +160,9 @@ class TestOTLPMetricsEncoder(unittest.TestCase):
                                                 start_time_unix_nano=1641946015139533244,
                                                 time_unix_nano=1641946016139533244,
                                                 as_int=33,
+                                                flags=int(
+                                                    DataPointFlags.get_default()
+                                                ),
                                             )
                                         ],
                                         aggregation_temporality=AggregationTemporality.CUMULATIVE,
@@ -504,6 +515,9 @@ class TestOTLPMetricsEncoder(unittest.TestCase):
                                                 ],
                                                 max=18.0,
                                                 min=8.0,
+                                                flags=int(
+                                                    DataPointFlags.get_default()
+                                                ),
                                             )
                                         ],
                                         aggregation_temporality=AggregationTemporality.DELTA,
@@ -635,6 +649,9 @@ class TestOTLPMetricsEncoder(unittest.TestCase):
                                                 ],
                                                 max=18.0,
                                                 min=8.0,
+                                                flags=int(
+                                                    DataPointFlags.get_default()
+                                                ),
                                             )
                                         ],
                                         aggregation_temporality=AggregationTemporality.DELTA,
@@ -697,6 +714,9 @@ class TestOTLPMetricsEncoder(unittest.TestCase):
                                                 ],
                                                 max=18.0,
                                                 min=8.0,
+                                                flags=int(
+                                                    DataPointFlags.get_default()
+                                                ),
                                             )
                                         ],
                                         aggregation_temporality=AggregationTemporality.DELTA,
@@ -767,6 +787,9 @@ class TestOTLPMetricsEncoder(unittest.TestCase):
                                                 ],
                                                 max=18.0,
                                                 min=8.0,
+                                                flags=int(
+                                                    DataPointFlags.get_default()
+                                                ),
                                             )
                                         ],
                                         aggregation_temporality=AggregationTemporality.DELTA,
@@ -837,6 +860,9 @@ class TestOTLPMetricsEncoder(unittest.TestCase):
                                                 ],
                                                 max=18.0,
                                                 min=8.0,
+                                                flags=int(
+                                                    DataPointFlags.get_default()
+                                                ),
                                             )
                                         ],
                                         aggregation_temporality=AggregationTemporality.DELTA,
@@ -868,7 +894,7 @@ class TestOTLPMetricsEncoder(unittest.TestCase):
                         zero_count=5,
                         positive=Buckets(offset=6, bucket_counts=[7, 8]),
                         negative=Buckets(offset=9, bucket_counts=[10, 11]),
-                        flags=12,
+                        flags=DataPointFlags.get_default(),
                         min=13.0,
                         max=14.0,
                     )
@@ -953,7 +979,9 @@ class TestOTLPMetricsEncoder(unittest.TestCase):
                                                     offset=9,
                                                     bucket_counts=[10, 11],
                                                 ),
-                                                flags=12,
+                                                flags=int(
+                                                    DataPointFlags.get_default()
+                                                ),
                                                 exemplars=[],
                                                 min=13.0,
                                                 max=14.0,