+++ opentelemetry-sdk/src/opentelemetry/sdk/metrics/export/__init__.py
@@ -17,6 +17,7 @@ from opentelemetry.sdk.metrics._internal.export import (
 # The point module is not in the export directory to avoid a circular import.
 from opentelemetry.sdk.metrics._internal.point import (  # noqa: F401
     Buckets,
+    DataPointFlags,
     DataPointT,
     DataT,
     ExponentialHistogram,
@@ -41,6 +42,7 @@ __all__ = [
     "MetricExportResult",
     "MetricReader",
     "PeriodicExportingMetricReader",
+    "DataPointFlags",
     "DataPointT",
     "DataT",
     "ExponentialHistogram",